Displaying query data
server.py: adjusted route /  to call a query, store it in a variable called 'friends', and created another variable that will be rendered with index.html called 'all_friends' that will hold the results of the query.
index.html:  Removed hard coded information and replaced with a for loop using jinja2 to loop through the array of dictionaries given from the all_friends variable and display the information accordingly.