Reset/Complete
After research and a few hours of testing.
Add a friend and remove a friend work with no issues.
Auto increment can be reset when deleting friends.
If ID count reaches 0/None, if else statements are use to avoid errors.
Will return if ahead in studies to add extras: Modify friends and delete specfic friend.